Adding ordered write option to PravegaStreamMirroringJob, reorganizing Main and AppConfiguration to client package
