+++ src/main/java/io/pravega/dataimporter/actions/KafkaMirroringAction.java
@@ -1,6 +1,6 @@
 package io.pravega.dataimporter.actions;
 
-import io.pravega.dataimporter.AppConfiguration;
+import io.pravega.dataimporter.client.AppConfiguration;
 import io.pravega.dataimporter.jobs.KafkaToPravegaStreamJob;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
