+++ src/main/java/io/pravega/dataimporter/client/AppConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.pravega.dataimporter;
+package io.pravega.dataimporter.client;
 
 import io.pravega.client.stream.ScalingPolicy;
 import io.pravega.client.stream.Stream;
